MAINT: Refactor example with Correlations
