+++ src/porepy/examples/geothermal_flow_via_correlations/.gitignore
@@ -1,0 +1,1 @@
+*.vtk
